move the app colors to colors package under theme package.
