--- app/src/main/java/com/amsterdam/cutetudee/presentation/theme/LightThemeColors.kt
@@ -1,2 +1,0 @@
-package com.amsterdam.cutetudee.presentation.theme
-
