+++ app/src/main/java/com/amsterdam/cutetudee/presentation/theme/colors/LightThemeColors.kt
@@ -1,0 +1,2 @@
+package com.amsterdam.cutetudee.presentation.theme.colors
+
